WindowManagerでアイテムウィンドウを管理できるように仕組みを大幅修正
