+++ ChernoRogueLikeUnityLast/Assets/Resources/Prefabs/Scene/Item/ItemCell.prefab.meta
