+++ ChernoRogueLikeUnityLast/Assets/Resources/Prefabs/Scene/Item/ItemWindow.prefab.meta
@@ -1,8 +1,8 @@
 fileFormatVersion: 2
-guid: 001bfe24990be3249a36acf692147fad
-folderAsset: yes
-DefaultImporter:
+guid: d493567c151481940be192320da3b9f1
+NativeFormatImporter:
   externalObjects: {}
+  mainObjectFileID: 100100000
   userData: 
   assetBundleName: 
   assetBundleVariant:
