+++ ChernoRogueLikeUnityLast/Assets/Scripts/Scenes/Item/ItemSample.cs.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: 921af31992f03e54f8fb262291aa2fbb
+guid: bd2391a24558e9a4aa7080f2c6a53f1a
 MonoImporter:
   externalObjects: {}
   serializedVersion: 2
