--- ChernoRogueLikeUnityLast/Assets/Scripts/Scenes/Item/SampleItem.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: a2f6b542814efb54d8ae1c085e357782
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
